// types/company.ts
export interface Company {
  id?: string;
  company_name?: string;
  company_domain?: string;
  company_linkedin_url?: string;
  company_website?: string;
  company_phone?: string;
  employee_count?: number;
  industry?: string;
  location?: string;
  logo_url?: string;
  leads_count?: number;
  jobs_count?: number;
  contacts_count?: number;
}